refactor(bucket): naming improvements
