+++ stacks/api/bucket/hooks/hook.spec.ts
@@ -1,6 +1,6 @@
 import {TestingModule, Test} from "@nestjs/testing";
 import {hookModuleProviders, SCHEMA} from "@spica-server/bucket/hooks/hook.module";
-import {ServicesModule} from "@spica-server/bucket/services/bucket.service.module";
+import {ServicesModule} from "@spica-server/bucket/services";
 import {DatabaseTestingModule} from "@spica-server/database/testing";
 import {PreferenceModule} from "@spica-server/preference";
 import {SchemaModule} from "@spica-server/core/schema";
