+++ stacks/api/bucket/services/index.ts
@@ -1,3 +1,3 @@
 export {BucketService} from "./bucket.service";
-export {ServicesModule} from "./bucket.service.module";
+export {ServicesModule} from "./services.module";
 export * from "./bucket";
